requirements.txt
openai
pydantic
python-dotenv
networkx
